simply move to vm package
