+++ security/cmd/node_agent/main.go
@@ -24,12 +24,12 @@ import (
 	"istio.io/istio/pkg/collateral"
 	"istio.io/istio/pkg/log"
 	"istio.io/istio/pkg/version"
-	"istio.io/istio/security/cmd/node_agent/na"
 	"istio.io/istio/security/pkg/cmd"
+	nvm "istio.io/istio/security/pkg/nodeagent/vm"
 )
 
 var (
-	naConfig = na.NewConfig()
+	naConfig = nvm.NewConfig()
 
 	rootCmd = &cobra.Command{
 		Use:   "node_agent",
@@ -94,7 +94,7 @@ func runNodeAgent() {
 		log.Errora(err)
 		os.Exit(-1)
 	}
-	nodeAgent, err := na.NewNodeAgent(naConfig)
+	nodeAgent, err := nvm.NewNodeAgent(naConfig)
 	if err != nil {
 		log.Errora(err)
 		os.Exit(-1)
