+++ security/cmd/node_agent_k8s/main.go
@@ -23,12 +23,12 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"istio.io/istio/security/cmd/node_agent/na"
 	"istio.io/istio/security/cmd/node_agent_k8s/workload/handler"
 	wlapi "istio.io/istio/security/cmd/node_agent_k8s/workloadapi"
 	"istio.io/istio/security/pkg/caclient"
 	"istio.io/istio/security/pkg/caclient/protocol"
 	"istio.io/istio/security/pkg/nodeagent/registry"
+	nvm "istio.io/istio/security/pkg/nodeagent/vm"
 	"istio.io/istio/security/pkg/platform"
 )
 
@@ -60,7 +60,7 @@ var (
 		Long:  "Node agent with both mgmt and workload api interfaces.",
 	}
 
-	naConfig na.Config
+	naConfig nvm.Config
 )
 
 func init() {
