+++ security/pkg/nodeagent/vm/nodeagent.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package na
+package vm
 
 import (
 	"fmt"
